requirements.txt
streamlit 
google-generativeai
googlesearch-python
pdfkit
